import buildImageTag
